Move /verify route registration to HubConnection
